move Overview and ThemedView into new dir '/views'
